wrlinux-images: fix for IMAGE_INSTALL in templates
Issue: LIN9-1812

The "+=" operations before "=" will be overridden, we used
process_templates.sh to generate extra image files for default
image and ensure that all "+=" operations of IMAGE_INSTALL in
templates are after the "=" operation defined in the original
image recipe.

But we don't have this mechanism after changing to setup.sh,
we can either:
- Adjust the order between IMAGE_INSTALL definition and
  "inherit wrlinux-image" in all image recipes just like
  the way used in the images in oe-core.
or
- Use _append instead of += in all templates and have this
  documented to avoid users trying to use += for IMAGE_INSTALL
  in local.conf or their own layers

The first option is chosen here.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>