Require PAM for vlock
Add a dependency on vlock to libpam, and update the libpam recipe
so that it fails if DISTRO_FEATURES does not contain pam, instead
of issuing a warning.

The vlock configure script requires that PAM either be disabled
or present.  But disabling it causes problems with bitbake,
unless pam is included in DISTRO_FEATURES.

Issue: LIN9-2196

Signed-off-by: T.O. Radzy Radzykewycz <radzy@windriver.com>