wr-base: add git version 2.11.0-rc2
Issue: LIN9-2134

We need our project to finish setup in about 6 minutes. This requires
git 2.11. So we add a new git recipe in wr-base. The version is 2.11.0-rc2.

The git.inc file is basically a copy from oe-core, with only a little
modification to suite the new version which uses git source instead of
tarball.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>