agent-proxy: drop the inc
Issue: TA164731

There is only one version, the inc is not needed
any more, drop the inc file and move contents
to the bb and drop INC_PR

Drop the DEFAULT_PREFERENCE as well which is used
for development version when there are more than
two versions.

The "jjj" in the HOMEPAGE should be a typo, remove it.

Signed-off-by: Jackie Huang <jackie.huang@windriver.com>