util-linux: disable schedutils for native
Issue: LIN9-2504

util-linux assumes glibc and kernel match. In specific, glibc shouldn't
declare that it supports SYS_sched_setattr which the kernel actually not.
This is true on most sane systems. But on some systems, this is not true.
And on such systems, when building util-linux-native, we would have the
following error.

| ../util-linux-2.28.1/schedutils/chrt.c:88:17: error: __NR_sched_setattr undeclared (first use in this function)
| return syscall(SYS_sched_setattr, pid, attr, flags);
| ^

Although we could modify chrt.c source and make it do not do such assumption,
I think a safer fix is to just use '--disable-schedutils' for util-linux-native.
And this patch does this.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>