autoconf-archive: move the new version from intel-x86 layer
Issue: LIN9-3864, LIN9-4167

The version of autoconf-archive in intel-x86 layer is 2016.09.16, and
the version in meta-oe is 2012.04.07.

We use the high version in intel-x86 bsp project with setting the
PREFERRED_VERSION in intel-x86.inc, but due to the priority of layers
(BBFILE_PRIORITY) between intel-x86 and meta-oe, the high version is
always compiled.
 BBFILE_PRIORITY_intel-x86 = "7"
 BBFILE_PRIORITY_openembedded-layer = "6"

At the same time, the variable of DEFAULT_PREFERENCE is -1 in high version
bbfile.

Now we move the new version into wr-base layer
 BBFILE_PRIORITY_wr-userspace-base = "6"

So the low version will be built in the other board project unless the
PREFERRED_VERSION is defined.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>