wrlinux-whitelist.conf: create initial virtualization-layer whitelist
LIN9 was released with a minimal copy of the meta-virtualization layer
in the wr-kernel repo. With efforts now underway to integratate
virtualization support into LIN9 we are about to drop this minimal
copy of meta-virtualization and instead use a full clone of the
upstream meta-virtualization repository.

Setup the initial whitelist to allow building the 7 recipes that are
present in the minimal copy. Once the full meta-virtualization is in
place we can expand this whitelist as needed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>